refactor!: 逻辑尺寸全链路 f32 化 + scale 字段删除 + 窗口尺寸族逻辑像素语义统一
- 逻辑宽高缓存 u32→f32，消除高 DPI/非整数 scale 下截断（1280/1.5=853.33→853）
- dpi.rs 删除 logical_size（与 to_pixel_size 重复），7 处调用点内联
- 删除 VireoWindow.scale 冗余字段（与 layout_scale() 恒等），drift 判定改用 layout_scale()
- Renderer::new/update_layout/resize 逻辑尺寸签名改 f32
- 删除已废弃的 WindowMetrics（prelude 导出移除；layout_size/layout_scale 取代）
- 直通 winit 的 16 个状态查询 getter 补 macOS exec_sync 阻塞/死锁警告文档；
  缓存 getter（layout_scale/layout_size/mouse_pos/focused/cursor_inside）注明不阻塞

+++ src/offscreen.rs
@@ -32,7 +32,7 @@ impl OffscreenCanvas {
         );
         let texture = Texture::new(&gpu.device, width, height, gpu.surface_format(),
             &gpu.texture_bind_group_layout, &gpu.default_sampler);
-        let renderer = Renderer::new(gpu.clone(), width, height, width, height, 1.0, aa, 1.0);
+        let renderer = Renderer::new(gpu.clone(), width as f32, height as f32, width, height, 1.0, aa, 1.0);
         Self { texture, renderer, init_duration, index: OffscreenIndex(0) }
     }
